feat: try to reduce event page first exec time
- Split and lazy-load event page mobile and desktop layouts
- Remove auto revalidation from event report form API call
- Lazy-render below-the-fold mobile items when visible
